requirements.txt
streamlit==1.25.0
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.24.3
